feat(auth): final remarks and add icons
