+++ client/src/App.js
@@ -1,12 +1,9 @@
 import React, { useState } from 'react'
-import { AuthPage } from './pages/AuthorizationPage/components'
+import { AuthPage } from './pages/Authorization'
 import { BlurBall } from './samples/BlurBallSample'
-
-import { usePush } from './hooks/push.hook'
 
 export const App = () => {
   const [login] = useState(false)
-  const push = usePush()
 
   return (
     <>
@@ -15,7 +12,6 @@ export const App = () => {
           login ? <div>Ты вошёл, нахуй пошёл</div> : <AuthPage />
         }
       </div>
-      <button onClick={() => push(['Kekglfadkjglksfdjgslkfdjglskdfjglksdfjglksdfjglksdfjgklsdfj', 'Lol'], true)}>Push mess</button>
       <BlurBall />
     </>
   )
