+++ client/src/pages/Authorization/ForgotPass/index.js
@@ -1,12 +1,16 @@
 import React from "react";
-import { TextInput } from "../../../../samples/TextInput";
-import st from "./../../../../styles/AuthPage/auth_page.module.css"
+import { TextInput } from "../../../samples/TextInput";
+import st from "../../../styles/AuthPage/auth_page.module.css"
 
+import emailIcon from '../../../assets/auth/email_icon.svg'
 
 export const ForgotPass = ({ form, setShowLogin, changeHandler }) => {
     return (
-        <div className={st.forgot_pass_container}>
-            <p className={st.go_back} onClick={setShowLogin.bind(this, true)}> X Выслать пароль</p>
+        <div className={st.auth_container}>
+            <div className={st.go_back} onClick={setShowLogin.bind(this, true)}>
+                <div>&times;</div>
+                <div>Выслать пароль</div>
+            </div>
             <TextInput
                 name='email'
                 value={form.email}
@@ -14,6 +18,7 @@ export const ForgotPass = ({ form, setShowLogin, changeHandler }) => {
                 topRound
                 bottomRound
                 placeholder='Email'
+                icon={emailIcon}
             />
             <div className={st.inbutton}>
                 <button>Выслать пароль</button>
