+++ client/src/styles/AuthPage/auth_page.module.css
@@ -6,8 +6,8 @@
     margin-top: 5%;
 }
 
-.logo {
-    height: 90px;
+.logo{
+    width: 90px;
 }
 
 .text {
@@ -21,29 +21,44 @@
 
 .go_back {
     position: absolute;
-    left: 15px;
-    top: 5px;
-    cursor: pointer;
+    left: 2%;
+    top: 2%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
     font-size: 18px;
-    font-weight: 800;
     color: #19344D;
+    cursor: pointer;
+}
+
+.go_back div:nth-child(1){
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding-top: 5px;
+    width: 40px;
+    height: 40px;
+    font-size: 40px;
+    font-weight: 200;
+}
+
+.go_back div:nth-child(2){
+    font-weight: 600;
 }
 
 .forgot {
-    float: right;
     cursor: pointer;
-    margin-top: 20px;
     color: #2090FB;
     font-size: 18px;
     text-decoration: underline;
 }
 
 .inbutton button {
+    width: 100%;
     margin-top: 20px;
     background: #2090FB;
     color: white;
     height: 50px;
-    width: 400px;
     border: none;
     border-radius: 10px;
     font-size: 18px;
@@ -52,4 +67,11 @@
 
 .inbutton button:active {
     background-color: #19344D;
+}
+
+.auth_options{
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    margin-top: 15px;
 }
