+++ client/src/styles/text_input.module.css
@@ -1,5 +1,8 @@
 .input_container{
-    padding: 15px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 15px 20px;
     height: 60px;
     background: #FFF;
     box-shadow: 0px 4px 40px rgba(50, 130, 207, 0.06);
@@ -7,11 +10,22 @@
 
 .input_container input{
     height: 100%;
-    width: 100%;
+    width: 95%;
     font-size: 18px;
     border: none;
 }
 
 .input_container input:focus{
     outline: none;
+}
+
+.icon_container{
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 5%;
+}
+
+.icon_container img{
+    width: 20px;
 }
